Struct libimagrt::runtime::Runtime

pub struct Runtime<'a> {
    // some fields omitted
}

Methods

impl<'a> Runtime<'a>

fn new(cli_spec: App<'a, 'a>) -> Result<Runtime<'a>, RuntimeError>

Gets the CLI spec for the program and retreives the config file path (or uses the default on in $HOME/.imag/config, $XDG_CONFIG_DIR/imag/config or from env("$IMAG_CONFIG") and builds the Runtime object with it.

The cli_spec object should be initially build with the ::get_default_cli_builder() function.

fn get_default_cli_builder(appname: &'a str, version: &'a str, about: &'a str) -> App<'a, 'a>

Get a commandline-interface builder object from clap

This commandline interface builder object already contains some predefined interface flags: * -v | --verbose for verbosity * --debug for debugging * -c | --config for alternative configuration file * -r | --rtp for alternative runtimepath * --store for alternative store path Each has the appropriate help text included.

The appname shall be "imag-".

fn arg_names() -> Vec<&'static str>

fn arg_verbosity_name() -> &'static str

fn arg_debugging_name() -> &'static str

fn arg_no_color_output_name() -> &'static str

fn arg_config_name() -> &'static str

fn arg_config_override_name() -> &'static str

fn arg_runtimepath_name() -> &'static str

fn arg_storepath_name() -> &'static str

fn arg_editor_name() -> &'static str

fn is_verbose(&self) -> bool

Get the verbosity flag value

fn is_debugging(&self) -> bool

Get the debugging flag value

fn rtp(&self) -> &PathBuf

Get the runtimepath

fn cli(&self) -> &ArgMatches

Get the commandline interface matches

fn config(&self) -> Option<&Configuration>

Get the configuration object

fn store(&self) -> &Store

Get the store object

fn editor(&self) -> Option<Command>